+++ cms/static/js/hera/store/actionTypes.jsx
@@ -1,0 +1,1 @@
+export const TITLE_CHANGED = 'TITLE_CHANGED';
