+++ cms/static/js/hera/store/reducers/index.jsx
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import TitleReducer from './TitleReducer';
+
+var reducers = combineReducers({
+    title: TitleReducer
+});
+
+export default reducers;
